--- utils/integration-tests/.env
@@ -1,1 +1,0 @@
-COMPOSE_PROJECT_NAME=directory-connector
